Depthmap/requirements.txt
h5py 
pyvista 
scikit-image 
scipy 
torchvision 
torch 
pillow
vtk